requirements.txt
numpy==1.21.0
scipy==1.7.0
matplotlib==3.4.3
pandas==1.3.3
tensorflow==2.6.0
torch==1.9.0
scikit-learn==0.24.2
pyyaml==5.4.1
jupyter==1.0.0
seaborn==0.11.2